Organize sidebar files under collections/components
