+++ frontend/src/metabase/collections/components/CollectionSidebar/CollectionSidebar.styled.jsx
