+++ frontend/src/metabase/collections/components/CollectionSidebar/CollectionSidebarFooter/CollectionSidebarFooter.jsx
