+++ frontend/src/metabase/collections/components/CollectionSidebar/CollectionSidebarFooter/CollectionSidebarFooter.styled.jsx
