+++ frontend/src/metabase/collections/components/CollectionSidebar/CollectionSidebarFooter/CollectionSidebarFooter.unit.spec.js
