+++ frontend/src/metabase/collections/components/CollectionSidebar/CollectionSidebarFooter/index.jsx
