+++ frontend/src/metabase/collections/components/CollectionSidebar/CollectionSidebarLink/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CollectionSidebarLink";
